// src/typings/utils.ts
export enum Category {
    music = "music",
    info = "info",
    dev = "dev",
    filters = "filters",
    playlist = "playlist",
    admin = "admin",
}

export enum Mode {
    production = "production",
    development = "development",
    test = "test",
}

export enum Language {
    EnglishUS = "EnglishUS",
    Spanish = "Spanish",
    French = "French", 
    German = "German",
    Indonesian = "Indonesian",
    Japanese = "Japanese",
    Korean = "Korean",
    Vietnamese = "Vietnamese",
    Arabic = "Arabic",
    Hindi = "Hindi",
    Punjabi = "Punjabi",
    Portuguese = "Portuguese",
    Russian = "Russian",
    Italian = "Italian",
    Turkish = "Turkish",
}